# specs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Scopegate API contract",
    "version": "1.0.0",
    "description": "Executable local product contract. Authority and trust boundaries are defined in docs/SECURITY.md."
  },
  "servers": [
    {
      "url": "http://localhost:5187",
      "description": "Local same-origin workspace"
    },
    {
      "url": "http://localhost:8457",
      "description": "Local API transport"
    }
  ],
  "paths": {
...
    "/auth/login": {
      "get": {
        "operationId": "startLogin",
        "summary": "Start OIDC authorization code flow with PKCE, state, and nonce",
        "security": [],
        "responses": {
          "302": {
            "description": "Redirect to configured identity provider"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "408": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/auth/callback": {
      "get": {
        "operationId": "finishLogin",
        "summary": "Verify provider callback and create an opaque session",
        "security": [],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Set secure session cookie and redirect to approved local destination"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "408": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "operationId": "getCurrentUser",
        "summary": "getCurrentUser",
        "tags": [
          "session"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "description": "Rate limit or temporary unavailability; retry only within the operation budget.",
            "headers": {
              "Retry-After": {
                "description": "Suggested delay in seconds when applicable.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Rate limit or temporary unavailability; retry only within the operation budget.",
            "headers": {
              "Retry-After": {
                "description": "Suggested delay in seconds when applicable.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "408": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/logout": {
      "post": {
        "operationId": "logout",
        "summary": "logout",
        "tags": [
          "session"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "description": "Rate limit or temporary unavailability; retry only within the operation budget.",
            "headers": {
              "Retry-After": {
                "description": "Suggested delay in seconds when applicable.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Rate limit or temporary unavailability; retry only within the operation budget.",
            "headers": {
              "Retry-After": {
                "description": "Suggested delay in seconds when applicable.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "204": {
            "description": "Session revoked and cookies removed."
          },
          "408": {
            "$ref": "#/components/responses/Error"
          },
          "413": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "scopegate_session"
      },
      "publisherBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Validated publisher machine credential; catalog write only"
      },
      "platformBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Validated platform operations credential; administrative provisioning only, no implicit resource use"
      }
    },
    "parameters": {
...
      "CsrfToken": {
        "name": "X-CSRF-Token",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 32
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string",
                "minLength": 1
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "correlation_id": {
                "type": "string",
                "format": "uuid"
              },
              "fields": {
                "type": "object",
                "additionalProperties": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "required": [
              "code",
              "message",
              "correlation_id"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "User": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "csrf_token": {
            "type": "string",
            "minLength": 32
          },
          "memberships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Membership"
            }
          },
          "email_verified": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "csrf_token",
          "memberships",
          "email_verified"
        ]
      },
...
      "Membership": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "type": "string",
            "enum": [
              "viewer",
              "access_manager"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "suspended"
            ]
          },
          "kind": {
            "type": "string",
            "enum": [
              "customer",
              "staff"
            ]
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "access_version": {
            "type": "integer",
            "minimum": 1
          },
          "can_review_migration": {
            "type": "boolean"
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "organization_name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "organization_id",
          "user_id",
          "role",
          "status",
          "kind",
          "access_version",
          "email"
        ]
      },
...
    },
    "responses": {
      "Error": {
        "description": "Stable machine-readable error. Inaccessible objects use 404; known insufficient management scope uses 403.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}